Legacy unit chart
