+++ src/modules/units/components/Units.tsx
@@ -3,12 +3,14 @@ import { Masonry } from '@mui/lab';
 import { Button, Paper, Stack } from '@mui/material';
 import { useCallback, useMemo, useState } from 'react';
 
+import CommonCard from '@/components/elements/CommonCard';
 import Loading from '@/components/elements/Loading';
 import PageTitle from '@/components/layout/PageTitle';
 import NotFound from '@/components/pages/NotFound';
 import { ProjectPermissionsFilter } from '@/modules/permissions/PermissionsFilters';
 import { useProjectDashboardContext } from '@/modules/projects/components/ProjectDashboard';
 import ProjectUnitsTable from '@/modules/units/components/ProjectUnitsTable';
+import UnitCapacityTable from '@/modules/units/components/UnitCapacityTable';
 import UnitGroupCard from '@/modules/units/components/UnitGroupCard';
 import UnitGroupFormDialog from '@/modules/units/components/UnitGroupFormDialog';
 import { ProjectDashboardRoutes } from '@/routes/routes';
@@ -84,6 +86,12 @@ const Units = () => {
               />
             ))}
           </Masonry>
+        )}
+        {/* keep legacy behavior: if units exist outside of groups, show utilization charts grouped by unit type */}
+        {unitGroups.length === 0 && (
+          <CommonCard title='Capacity'>
+            <UnitCapacityTable projectId={project.id} />
+          </CommonCard>
         )}
         <Paper>
           <ProjectUnitsTable projectId={project.id} />
